deque braught into tracer
