+++ jac/jaclang/runtimelib/gins/cfg.py
@@ -267,7 +267,6 @@ def find_block_by_offset(block_map: BlockMap, offset: int) -> int:
     return None
 
 
-# Function to visualize CFG using Graphviz
 def visualize_cfg(cfg: CFG):
     dot = Digraph(comment="Control Flow Graph")
     for node in cfg.nodes:
@@ -278,21 +277,23 @@ def visualize_cfg(cfg: CFG):
     return dot
 
 
-if __name__ == "__main__":
-    # Sample list of instructions for processing
-    ##simple=
-    # instructions = disassemble_bytecode(b'c\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x02\x00\x00\x00\x00\x00\x00\x01\xf3*\x00\x00\x00\x97\x00d\x00d\x01l\x00m\x01Z\x01\x01\x00d\x00Z\x02d\x00Z\x03e\x03d\x00k\\\x00\x00r\x02d\x02Z\x02d\x03Z\x02y\x04)\x05\xe9\x00\x00\x00\x00)\x01\xda\x0bannotations\xe9\x01\x00\x00\x00\xe9\xff\xff\xff\xffN)\x04\xda\n__future__r\x02\x00\x00\x00\xda\x01a\xda\x01x\xa9\x00\xf3\x00\x00\x00\x00\xfaP/Users/jakobtherkelsen/Documents/jaseci-ginS/jac/examples/ginsScripts/simple.jac\xfa\x08<module>r\x0b\x00\x00\x00\x01\x00\x00\x00s%\x00\x00\x00\xf0\x03\x01\x01\x01\xf5\x02\x07\x02\x03\xd8\x05\x06\x801\xd8\x05\x06\x801\xd8\x06\x07\x881\x82f\xd8\x07\x08\x80Q\xe0\x05\x07\x811r\t\x00\x00\x00')
-    # hot path
-    instructions = disassemble_bytecode(
-        b"c\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x04\x00\x00\x00\x00\x00\x00\x01\xf3\xa4\x00\x00\x00\x97\x00U\x00d\x00d\x01l\x00m\x01Z\x01\x01\x00d\x00Z\x02d\x02e\x03d\x03<\x00\x00\x00d\x04Z\x04d\x02e\x03d\x05<\x00\x00\x00e\x02e\x04z\x00\x00\x00Z\x05d\x02e\x03d\x06<\x00\x00\x00\x02\x00e\x06d\x04\xab\x01\x00\x00\x00\x00\x00\x00D\x00]\x1c\x00\x00Z\x07e\x02d\x07z\x06\x00\x00r\ne\x04e\x02z\x00\x00\x00Z\x04e\x02e\x04z\x00\x00\x00Z\x05d\x08e\x07z\x05\x00\x00e\x04e\x05z\x05\x00\x00z\x00\x00\x00Z\x02\x8c\x1e\x04\x00d\tZ\x05e\x05d\tk(\x00\x00r\x03d\x08Z\x05y\ny\n)\x0b\xe9\x00\x00\x00\x00)\x01\xda\x0bannotations\xda\x03int\xda\x01x\xe9\x03\x00\x00\x00\xda\x01y\xda\x01z\xe9\x02\x00\x00\x00\xe9\x04\x00\x00\x00\xe9\n\x00\x00\x00N)\x08\xda\n__future__r\x02\x00\x00\x00r\x04\x00\x00\x00\xda\x0f__annotations__r\x06\x00\x00\x00r\x07\x00\x00\x00\xda\x05range\xda\x01i\xa9\x00\xf3\x00\x00\x00\x00\xfaU/Users/jakobtherkelsen/Documents/jaseci-ginS/jac/examples/gins_scripts/simple_for.jac\xda\x08<module>r\x12\x00\x00\x00\x01\x00\x00\x00sy\x00\x00\x00\xf0\x03\x01\x01\x01\xf6\x02\x0f\x02\x03\xd8\r\x0e\x80Q\x80s\x83Z\xd8\r\x0e\x80Q\x80s\x83Z\xd8\r\x0e\x90\x11\x89U\x80Q\x80s\x83^\xd9\x0e\x13\x90A\x8eh\x88\x11\xd8\n\x0b\x88a\x8a%\xd8\r\x0e\x90\x11\x89U\x88\x11\xd8\r\x0e\x90\x11\x89U\x88\x11\xe0\x0b\x0c\x88q\x895\x901\x98\x01\x917\x89?\x81q\xf0\x0b\x00\x0f\x17\xf0\x0e\x00\n\x0c\x80Q\xd8\x08\t\x88R\x8a\x07\xd8\n\x0b\x81q\xf0\x03\x00\t\x10r\x10\x00\x00\x00"
-    )
-    BBs = create_BBs(instructions)
-    print(BBs)
+# can use as a test
 
-    cfg = create_cfg(BBs)
-    print("\nControl Flow Graph (CFG):")
-    print(cfg)
+# if __name__ == "__main__":
+#     # Sample list of instructions for processing
+#     ##simple=
+#     # instructions = disassemble_bytecode(b'c\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x02\x00\x00\x00\x00\x00\x00\x01\xf3*\x00\x00\x00\x97\x00d\x00d\x01l\x00m\x01Z\x01\x01\x00d\x00Z\x02d\x00Z\x03e\x03d\x00k\\\x00\x00r\x02d\x02Z\x02d\x03Z\x02y\x04)\x05\xe9\x00\x00\x00\x00)\x01\xda\x0bannotations\xe9\x01\x00\x00\x00\xe9\xff\xff\xff\xffN)\x04\xda\n__future__r\x02\x00\x00\x00\xda\x01a\xda\x01x\xa9\x00\xf3\x00\x00\x00\x00\xfaP/Users/jakobtherkelsen/Documents/jaseci-ginS/jac/examples/ginsScripts/simple.jac\xfa\x08<module>r\x0b\x00\x00\x00\x01\x00\x00\x00s%\x00\x00\x00\xf0\x03\x01\x01\x01\xf5\x02\x07\x02\x03\xd8\x05\x06\x801\xd8\x05\x06\x801\xd8\x06\x07\x881\x82f\xd8\x07\x08\x80Q\xe0\x05\x07\x811r\t\x00\x00\x00')
+#     # hot path
+#     instructions = disassemble_bytecode(
+#         b"c\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x04\x00\x00\x00\x00\x00\x00\x01\xf3\xa4\x00\x00\x00\x97\x00U\x00d\x00d\x01l\x00m\x01Z\x01\x01\x00d\x00Z\x02d\x02e\x03d\x03<\x00\x00\x00d\x04Z\x04d\x02e\x03d\x05<\x00\x00\x00e\x02e\x04z\x00\x00\x00Z\x05d\x02e\x03d\x06<\x00\x00\x00\x02\x00e\x06d\x04\xab\x01\x00\x00\x00\x00\x00\x00D\x00]\x1c\x00\x00Z\x07e\x02d\x07z\x06\x00\x00r\ne\x04e\x02z\x00\x00\x00Z\x04e\x02e\x04z\x00\x00\x00Z\x05d\x08e\x07z\x05\x00\x00e\x04e\x05z\x05\x00\x00z\x00\x00\x00Z\x02\x8c\x1e\x04\x00d\tZ\x05e\x05d\tk(\x00\x00r\x03d\x08Z\x05y\ny\n)\x0b\xe9\x00\x00\x00\x00)\x01\xda\x0bannotations\xda\x03int\xda\x01x\xe9\x03\x00\x00\x00\xda\x01y\xda\x01z\xe9\x02\x00\x00\x00\xe9\x04\x00\x00\x00\xe9\n\x00\x00\x00N)\x08\xda\n__future__r\x02\x00\x00\x00r\x04\x00\x00\x00\xda\x0f__annotations__r\x06\x00\x00\x00r\x07\x00\x00\x00\xda\x05range\xda\x01i\xa9\x00\xf3\x00\x00\x00\x00\xfaU/Users/jakobtherkelsen/Documents/jaseci-ginS/jac/examples/gins_scripts/simple_for.jac\xda\x08<module>r\x12\x00\x00\x00\x01\x00\x00\x00sy\x00\x00\x00\xf0\x03\x01\x01\x01\xf6\x02\x0f\x02\x03\xd8\r\x0e\x80Q\x80s\x83Z\xd8\r\x0e\x80Q\x80s\x83Z\xd8\r\x0e\x90\x11\x89U\x80Q\x80s\x83^\xd9\x0e\x13\x90A\x8eh\x88\x11\xd8\n\x0b\x88a\x8a%\xd8\r\x0e\x90\x11\x89U\x88\x11\xd8\r\x0e\x90\x11\x89U\x88\x11\xe0\x0b\x0c\x88q\x895\x901\x98\x01\x917\x89?\x81q\xf0\x0b\x00\x0f\x17\xf0\x0e\x00\n\x0c\x80Q\xd8\x08\t\x88R\x8a\x07\xd8\n\x0b\x81q\xf0\x03\x00\t\x10r\x10\x00\x00\x00"
+#     )
+#     BBs = create_BBs(instructions)
+#     print(BBs)
 
-    # Visualize CFG
-    dot = visualize_cfg(cfg)
-    dot.render("cfg.gv", view=True)
+#     cfg = create_cfg(BBs)
+#     print("\nControl Flow Graph (CFG):")
+#     print(cfg)
+
+#     # Visualize CFG
+#     dot = visualize_cfg(cfg)
+#     dot.render("cfg.gv", view=True)
